Userdata manage

+++ src/Shopping/Server/Services/Implementations/EfDataRepository.cs
@@ -309,6 +309,9 @@ namespace Shopping.Server.Services.Implementations
         public async Task<bool> DeleteShoppingListAsync(string id)
         {
             var existing = await GetShoppingListAsync(id);
+
+            await RemoveAssignmentsOfShoppingListAsync(id);
+
             _context.ShoppingLists.Remove(existing);
 
             bool result = false;
@@ -384,6 +387,9 @@ namespace Shopping.Server.Services.Implementations
         public async Task<bool> DeleteUserGroupAsync(string id)
         {
             var existing = await GetUserGroupAsync(id);
+
+            await RemoveAssignmentsOfGroupAsync(id);
+
             _context.UserGroups.Remove(existing);
 
             bool result = false;
@@ -452,20 +458,28 @@ namespace Shopping.Server.Services.Implementations
 
             return existing;
         }
-        public async Task<bool> DeleteGroupListAssignmentAsync(string id)
+        public async Task<bool> DeleteGroupListAssignmentAsync(UserGroupShoppingList assignment)
         {
-            var existing = await GetGroupListAssignmentAsync(id);
-            _context.UserGroupShoppingLists.Remove(existing);
+            var existing = (await GetGroupListAssignmentsAsync())
+                .FirstOrDefault(a => a.ShoppingListId == assignment.ShoppingListId && a.UserGroupId == assignment.UserGroupId);
 
             bool result = false;
-            try
+            if (existing != null)
             {
-                await SaveChangesAsync();
-                result = true;
+                _context.UserGroupShoppingLists.Remove(existing);
+                try
+                {
+                    await SaveChangesAsync();
+                    result = true;
+                }
+                catch
+                {
+                    result = false;
+                }
             }
-            catch
+            else
             {
-                result = false;
+                _logger.LogWarning($"No assignment found between list '{assignment.ShoppingListId}' and group '{assignment.UserGroupId}'");
             }
 
             return result;
@@ -481,6 +495,33 @@ namespace Shopping.Server.Services.Implementations
             var itemsWithoutCurrent = assignments.Where(a => a.Id != item.Id).ToList();
 
             return !(itemsWithoutCurrent.Any(a => a.ShoppingListId == item.ShoppingListId && a.UserGroupId == item.UserGroupId));
+        }
+        private async Task<bool> RemoveAssignmentsOfGroupAsync(string userGroupId)
+        {
+            var allAssignmentsOfGroup = (await GetGroupListAssignmentsAsync())
+                .Where(a => a.UserGroupId == userGroupId)
+                .ToList();
+
+            return await DeleteAssignments(allAssignmentsOfGroup);
+        }
+        private async Task<bool> RemoveAssignmentsOfShoppingListAsync(string shoppingListId)
+        {
+            var allAssignmentsOfList = (await GetGroupListAssignmentsAsync())
+                .Where(a => a.ShoppingListId == shoppingListId)
+                .ToList();
+
+            return await DeleteAssignments(allAssignmentsOfList);
+        }
+        private async Task<bool> DeleteAssignments(List<UserGroupShoppingList> assignments)
+        {
+            foreach (var assignment in assignments)
+            {
+                if (!(await DeleteGroupListAssignmentAsync(assignment)))
+                {
+                    return false;
+                }
+            }
+            return true;
         }
         #endregion
